Segregated check functions.
